requirements.txt
# Requirements file for Connect Project
# Complete dependencies for Agent API, MCP Server, and Core Services

# Core Data Processing
pandas

# Database
neo4j

# Configuration
python-dotenv

# AI/ML Libraries
langchain
langchain-community
langchain-openai
langchain-neo4j
openai

# LangGraph Agent Dependencies
langgraph
langgraph-checkpoint
httpx
aiohttp

# FastAPI & Web Server (Agent API + MCP Server)
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart

# Security
passlib[bcrypt]

# Utilities
aiofiles

# Testing
pytest
pytest-asyncio

# Additional dependencies for production deployment
gunicorn
python-json-logger